test(AuthSteps): 로그인요청 메서드 응집화
